feat: adds `EagerExprNameNamespace`
Didn't realise until now that both backends did the same thing
