+++ narwhals/_arrow/expr.py
@@ -8,7 +8,6 @@ from typing import Sequence
 
 import pyarrow.compute as pc
 
-from narwhals._arrow.expr_name import ArrowExprNameNamespace
 from narwhals._arrow.series import ArrowSeries
 from narwhals._compliant import EagerExpr
 from narwhals._expression_parsing import ExprKind
@@ -263,7 +262,3 @@ class ArrowExpr(EagerExpr["ArrowDataFrame", ArrowSeries]):
         return self._reuse_series("rank", method=method, descending=descending)
 
     ewm_mean = not_implemented()
-
-    @property
-    def name(self: Self) -> ArrowExprNameNamespace:
-        return ArrowExprNameNamespace(self)
